Visualization/requirements.txt
# Function dependencies, for example:
# package>=version
boto3
pandas
google-cloud-storage
simplejson